Add dark mode to navigation controller
